Moving into graphql folder - makes more sense
