--- Source/typescript/backend/data/index.ts
@@ -1,8 +1,0 @@
-// Copyright (c) Dolittle. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-import './GuidExtensions';
-
-export * from './GuidExtensions';
-export * from './GuidScalar';
-export * from './GraphQLSchema';
