+++ Source/typescript/backend/graphql/GuidExtensions.ts
