+++ Source/typescript/backend/graphql/GuidScalar.ts
